requirements.txt
pytz==2024.2
requests==2.32.3
PyJWT>=2.0
cryptography>=3.4.8
python-dotenv==1.0.1
boto3==1.38.26